third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/coep-navigate-popup.https_4-last-expected.txt

This is a testharness.js-based test.
FAIL Popup navigating to same-origin with coop unsafe-none/coep assert_equals: Openee browsing context closed expected true but got false
FAIL Popup navigating to same-site with coop unsafe-none/coep assert_equals: Openee browsing context closed expected true but got false
FAIL Popup navigating to same-origin with coop unsafe-none/no coep assert_equals: Openee browsing context closed expected true but got false
FAIL Popup navigating to same-site with coop unsafe-none/no coep assert_equals: Openee browsing context closed expected true but got false
Harness: the test ran to completion.
